# SML runtime optimizations