# What does this implement/fix?

This PR includes some optimizations to the SML component.

While using the SML component to decode a number of SML messages from a smart-meter, the warning `Component sml took a long time for an operation (66ms)` appeared. This is with logger level `INFO`, with `DEBUG` the processing time increases to 134ms.

This was tested on an ESP32C3, which does not contain native floating-point support.

By doing a code review, the following issues were identified and fixed:

- Superfluous string formatting of debug messages when debug level is not enabled.
- Superfluous coping of byte vectors where a byte view is appropriate.
- Superfluous string formatting within a for loop.

With these optimizations, SML runtime changes as follows, depending on the logger level:

```code
DEBUG: 134ms -> 97ms
INFO: 66ms -> 25ms
```

## Types of changes

- [ ] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [x] Code quality improvements to existing code or addition of tests
- [ ] Other

**Related issue or feature (if applicable):**

- none

**Pull request in [esphome-docs](https://github.com/esphome/esphome-docs) with documentation (if applicable):**

- none

## Test Environment

- [x] ESP32
- [ ] ESP32 IDF
- [ ] ESP8266
- [ ] RP2040
- [ ] BK72xx
- [ ] RTL87xx

## Example entry for `config.yaml`:

```yaml
esphome:
  name: esphome-smart-meter
  friendly_name: "ESPHome Smart Meter"

esp32:
  board: seeed_xiao_esp32c3
  variant: esp32c3
  framework:
    type: arduino

wifi:
  ssid: !secret wifi_ssid
  password: !secret wifi_password

ota:
  - platform: esphome
    password: !secret ota_password

logger:
  level: INFO

api:
  encryption:
    key: !secret api_encryption_key

uart:
  id: uart_bus
  rx_pin: GPIO20 # D7
  baud_rate: 9600

sml:
  id: smart_meter
  uart_id: uart_bus

sensor:
  - platform: uptime
    name: "Uptime"
  - platform: wifi_signal
    name: "WiFi Signal"
  - platform: sml
    id: enery_consumed_total
    name: "Zählerstand Bezug"
    sml_id: smart_meter
    obis_code: "1-0:1.8.0"
    accuracy_decimals: 4
    unit_of_measurement: kWh
    device_class: energy
    state_class: total_increasing
    filters:
      - multiply: 0.0001
      - throttle_average: 30s
  - platform: sml
    id: enery_produced_total
    name: "Zählerstand Einspeisung"
    sml_id: smart_meter
    obis_code: "1-0:2.8.0"
    accuracy_decimals: 4
    unit_of_measurement: kWh
    device_class: energy
    state_class: total_increasing
    filters:
      - multiply: 0.0001
      - throttle_average: 30s
  - platform: sml
    id: power
    name: "Wirkleistung"
    sml_id: smart_meter
    obis_code: "1-0:16.7.0"
    accuracy_decimals: 0
    unit_of_measurement: W
    device_class: power
    state_class: measurement
  - platform: sml
    id: power_l1
    name: "Wirkleistung L1"
    sml_id: smart_meter
    obis_code: "1-0:36.7.0"
    accuracy_decimals: 0
    unit_of_measurement: W
    device_class: power
    state_class: measurement
  - platform: sml
    id: power_l2
    name: "Wirkleistung L2"
    sml_id: smart_meter
    obis_code: "1-0:56.7.0"
    accuracy_decimals: 0
    unit_of_measurement: W
    device_class: power
    state_class: measurement
  - platform: sml
    id: power_l3
    name: "Wirkleistung L3"
    sml_id: smart_meter
    obis_code: "1-0:76.7.0"
    accuracy_decimals: 0
    unit_of_measurement: W
    device_class: power
    state_class: measurement
  - platform: sml
    id: voltage_l1
    name: "Spannung L1"
    sml_id: smart_meter
    obis_code: "1-0:32.7.0"
    accuracy_decimals: 1
    unit_of_measurement: "V"
    device_class: voltage
    state_class: measurement
    filters:
      - multiply: 0.1
  - platform: sml
    id: voltage_l2
    name: "Spannung L2"
    sml_id: smart_meter
    obis_code: "1-0:52.7.0"
    accuracy_decimals: 1
    unit_of_measurement: "V"
    device_class: voltage
    state_class: measurement
    filters:
      - multiply: 0.1
  - platform: sml
    id: voltage_l3
    name: "Spannung L3"
    sml_id: smart_meter
    obis_code: "1-0:72.7.0"
    accuracy_decimals: 1
    unit_of_measurement: "V"
    device_class: voltage
    state_class: measurement
    filters:
      - multiply: 0.1
  - platform: sml
    id: current_l1
    name: "Strom L1"
    sml_id: smart_meter
    obis_code: "1-0:31.7.0"
    accuracy_decimals: 2
    unit_of_measurement: A
    device_class: current
    state_class: measurement
    filters:
      - multiply: 0.01
  - platform: sml
    id: current_l2
    name: "Strom L2"
    sml_id: smart_meter
    obis_code: "1-0:51.7.0"
    accuracy_decimals: 2
    unit_of_measurement: A
    device_class: current
    state_class: measurement
    filters:
      - multiply: 0.01
  - platform: sml
    id: current_l3
    name: "Strom L3"
    sml_id: smart_meter
    obis_code: "1-0:71.7.0"
    accuracy_decimals: 2
    unit_of_measurement: A
    device_class: current
    state_class: measurement
    filters:
      - multiply: 0.01
  - platform: sml
    id: phase_u2_u1
    name: "Phasenwinkel U2 zu U1"
    sml_id: smart_meter
    obis_code: "1-0:81.7.1"
    accuracy_decimals: 0
    unit_of_measurement: °
    device_class: power_factor
    state_class: measurement
  - platform: sml
    id: phase_u3_u1
    name: "Phasenwinkel U3 zu U1"
    sml_id: smart_meter
    obis_code: "1-0:81.7.2"
    accuracy_decimals: 0
    unit_of_measurement: °
    device_class: power_factor
    state_class: measurement
  - platform: sml
    id: phase_i1_u1
    name: "Phasenwinkel I1 zu U1"
    sml_id: smart_meter
    obis_code: "1-0:81.7.4"
    accuracy_decimals: 0
    unit_of_measurement: °
    device_class: power_factor
    state_class: measurement
  - platform: sml
    id: phase_i2_u2
    name: "Phasenwinkel I2 zu U2"
    sml_id: smart_meter
    obis_code: "1-0:81.7.15"
    accuracy_decimals: 0
    unit_of_measurement: °
    device_class: power_factor
    state_class: measurement
  - platform: sml
    id: phase_i3_u3
    name: "Phasenwinkel I3 zu U3"
    sml_id: smart_meter
    obis_code: "1-0:81.7.26"
    accuracy_decimals: 0
    unit_of_measurement: °
    device_class: power_factor
    state_class: measurement
  - platform: sml
    id: grid_frequency
    name: "Netzfrequenz"
    sml_id: smart_meter
    obis_code: "1-0:14.7.0"
    accuracy_decimals: 1
    unit_of_measurement: Hz
    device_class: frequency
    state_class: measurement
    filters:
      - multiply: 0.1
```

## Checklist:
  - [x] The code change is tested and works locally.
  - [ ] Tests have been added to verify that the new code works (under `tests/` folder).

If user exposed functionality or configuration variables are added/changed:
  - [ ] Documentation added/updated in [esphome-docs](https://github.com/esphome/esphome-docs).